requirements.txt
asgiref==3.5.2
colorama==0.4.6
cors==1.0.1
distro==1.7.0
Django==3.2.10
django-cors-headers==3.13.0
django-rest==0.8.7
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
filelock==3.13.1
future==0.18.3
gevent==23.9.1
greenlet==3.0.3
PyJWT==1.7.1
pytz==2022.1
setuptools==68.2.2
sqlparse==0.4.3
tldextract==5.1.1
wheel==0.41.2
zope.event==5.0
zope.interface==6.1
